Move Expected's to a separate package
