+++ testutil-server/src/main/java/io/spine/server/CommandHandlerTest.java
@@ -26,6 +26,7 @@ import io.spine.client.ActorRequestFactory;
 import io.spine.client.TestActorRequestFactory;
 import io.spine.core.Command;
 import io.spine.server.command.CommandHandlingEntity;
+import io.spine.server.expected.CommandExpected;
 import io.spine.server.model.HandlerMethodFailedException;
 
 import java.util.List;
