+++ testutil-server/src/test/java/io/spine/server/aggregate/given/AggregateCommandTestShouldEnv.java
@@ -24,7 +24,7 @@ import com.google.protobuf.Message;
 import com.google.protobuf.StringValue;
 import com.google.protobuf.Timestamp;
 import com.google.protobuf.util.Timestamps;
-import io.spine.server.CommandExpected;
+import io.spine.server.expected.CommandExpected;
 import io.spine.server.aggregate.Aggregate;
 import io.spine.server.aggregate.AggregateCommandTest;
 import io.spine.server.aggregate.AggregateRepository;
