+++ testutil-server/src/test/java/io/spine/server/aggregate/given/AggregatePartCommandTestShouldEnv.java
@@ -25,7 +25,7 @@ import com.google.protobuf.StringValue;
 import com.google.protobuf.UInt32Value;
 import io.spine.base.Identifier;
 import io.spine.server.BoundedContext;
-import io.spine.server.CommandExpected;
+import io.spine.server.expected.CommandExpected;
 import io.spine.server.aggregate.AggregatePart;
 import io.spine.server.aggregate.AggregatePartCommandTest;
 import io.spine.server.aggregate.AggregatePartRepository;
